Add axum-prometheus to expose metrics

+++ src/alarmservice-rs/src/api/handlers/rooms_handler.rs
@@ -41,5 +41,5 @@ pub async fn active_alarm_count_handler(
 
 #[derive(Serialize, Deserialize)]
 pub struct AlarmCountDto {
-    pub active_alarm_count: i32,
+    pub active_alarm_count: i64,
 }
